fix: handle case where we add NodeSelectorTerms to existing list (#31)
* feat: add nodeaffinity initialization from a kubeconfig file

This adds an optional input argument ``--kubeconfig` to the nodeaffinity main, which accepts a path to a local kube config file (for example, ~/.kube/config).  If specified, this makes the nodeaffinity injector use the kubeconfig file to connect to the kubernetes cluster.  The purpose of this is for local testing.

* fix: correctly handle case where we add NodeSelectorTerms to existing list

Prior to this fix, the injecting NodeSelectorTerms on a Pod that already had some specified would result in the new NodeSelectorTerms being nested one extra level in the array.  For example, running on this pod:

```
apiVersion: v1
kind: Pod
metadata:
  name: nginx
  namespace: kubeflow
spec:
  affinity:
    nodeAffinity:
      requiredDuringSchedulingIgnoredDuringExecution:
        nodeSelectorTerms:
        - matchExpressions:
          - key: existing-key1
            operator: In
            values:
            - existing-value1
  containers:
  - image: nginx:1.14.2
    name: nginx
    ports:
    - containerPort: 80
```

Yielded a patched version with affinity of:
```
spec:
  affinity:
    nodeAffinity:
      requiredDuringSchedulingIgnoredDuringExecution:
        nodeSelectorTerms:
        - matchExpressions:
          - key: existing-key1
            operator: In
            values:
            - existing-value1
        - - matchExpressions:   <---extra nesting
            - key: the-testing-key
              operator: In
              values:
              - the-testing-val1
          - matchExpressions:
            - key: the-testing-key2
              operator: In
              values:
              - the-testing-val2
```

This commit corrects this by refactoring the way patches are build such that:
* a patch is created to initialise an empty affinity.nodeAffinity.requiredDuringSchedulingIgnoredDuringExecution.nodeSelectorTerms array, if it does not already exist (including any parent objects required)
* N patches are created, one for each of the NodeSelectorTerms to be added (rather than a single patch that would replace the entire array)

* fix: re-enable injector TestMutate

This fixes and re-enables TestMutate to test Injector.mutate for the updated code.

* chore: clean up old comments, typos